+++ application-bemyplan/build.gradle
@@ -7,13 +7,13 @@ dependencies {
     api 'org.springframework.boot:spring-boot-starter-web'
     implementation 'org.springframework.boot:spring-boot-starter-data-redis'
     implementation 'org.springframework.session:spring-session-data-redis'
-    implementation "org.springframework.cloud:spring-cloud-starter-openfeign"
+    api "org.springframework.cloud:spring-cloud-starter-openfeign"
     implementation 'mysql:mysql-connector-java'
     implementation 'org.jetbrains:annotations:20.1.0'
     api group: 'io.springfox', name: 'springfox-boot-starter', version: '3.0.0'
-    implementation group: 'io.jsonwebtoken', name: 'jjwt-api', version: '0.11.2'
-    implementation group: 'io.jsonwebtoken', name: 'jjwt-impl', version: '0.11.2'
-    implementation group: 'io.jsonwebtoken', name: 'jjwt-jackson', version: '0.11.2'
+    api group: 'io.jsonwebtoken', name: 'jjwt-api', version: '0.11.2'
+    api group: 'io.jsonwebtoken', name: 'jjwt-impl', version: '0.11.2'
+    api group: 'io.jsonwebtoken', name: 'jjwt-jackson', version: '0.11.2'
     implementation 'org.mapstruct:mapstruct:1.4.2.Final'
     implementation 'org.projectlombok:lombok-mapstruct-binding:0.2.0'
