+++ applications/app-bemyplan/src/main/java/com/deploy/bemyplan/auth/remote/google/GoogleProfileInfoResponse.java
@@ -1,22 +1,23 @@
-package com.deploy.bemyplan.external.client.kakao.dto.response;
+package com.deploy.bemyplan.auth.remote.google;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
+import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.ToString;
 
 @ToString
 @Getter
-@NoArgsConstructor
+@NoArgsConstructor(access = AccessLevel.PRIVATE)
 @JsonIgnoreProperties(ignoreUnknown = true)
 @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
-public class KaKaoProfileResponse {
+public class GoogleProfileInfoResponse {
 
     private String id;
 
-    private KaKaoProfileResponse(String id) {
+    private GoogleProfileInfoResponse(String id) {
         this.id = id;
     }
 }
