+++ applications/app-bemyplan/src/main/java/com/deploy/bemyplan/auth/remote/kakao/KaKaoProfileResponse.java
@@ -1,23 +1,22 @@
-package com.deploy.bemyplan.external.client.google.dto.response;
+package com.deploy.bemyplan.auth.remote.kakao;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
-import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.ToString;
 
 @ToString
 @Getter
-@NoArgsConstructor(access = AccessLevel.PRIVATE)
+@NoArgsConstructor
 @JsonIgnoreProperties(ignoreUnknown = true)
 @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
-public class GoogleProfileInfoResponse {
+public class KaKaoProfileResponse {
 
     private String id;
 
-    private GoogleProfileInfoResponse(String id) {
+    private KaKaoProfileResponse(String id) {
         this.id = id;
     }
 }
