+++ applications/app-bemyplan/src/main/java/com/deploy/bemyplan/auth/service/impl/GoogleAuthService.java
@@ -1,12 +1,12 @@
 package com.deploy.bemyplan.auth.service.impl;
 
+import com.deploy.bemyplan.auth.remote.google.GoogleAuthApiClient;
+import com.deploy.bemyplan.auth.remote.google.GoogleProfileInfoResponse;
 import com.deploy.bemyplan.auth.service.AuthService;
 import com.deploy.bemyplan.auth.service.dto.LoginDto;
 import com.deploy.bemyplan.auth.service.dto.SignUpDto;
 import com.deploy.bemyplan.domain.user.UserRepository;
 import com.deploy.bemyplan.domain.user.UserSocialType;
-import com.deploy.bemyplan.external.client.google.GoogleAuthApiClient;
-import com.deploy.bemyplan.external.client.google.dto.response.GoogleProfileInfoResponse;
 import com.deploy.bemyplan.user.service.UserService;
 import com.deploy.bemyplan.user.service.UserServiceUtils;
 import lombok.RequiredArgsConstructor;
