+++ applications/app-bemyplan/src/main/java/com/deploy/bemyplan/auth/service/impl/KaKaoAuthService.java
@@ -1,12 +1,12 @@
 package com.deploy.bemyplan.auth.service.impl;
 
+import com.deploy.bemyplan.auth.remote.kakao.KaKaoAuthApiClient;
+import com.deploy.bemyplan.auth.remote.kakao.KaKaoProfileResponse;
 import com.deploy.bemyplan.auth.service.AuthService;
 import com.deploy.bemyplan.auth.service.dto.LoginDto;
 import com.deploy.bemyplan.auth.service.dto.SignUpDto;
 import com.deploy.bemyplan.domain.user.UserRepository;
 import com.deploy.bemyplan.domain.user.UserSocialType;
-import com.deploy.bemyplan.external.client.kakao.KaKaoAuthApiClient;
-import com.deploy.bemyplan.external.client.kakao.dto.response.KaKaoProfileResponse;
 import com.deploy.bemyplan.user.service.UserService;
 import com.deploy.bemyplan.user.service.UserServiceUtils;
 import lombok.RequiredArgsConstructor;
